mreza refaktorirana
