+++ suap_end_device/main/CMakeLists.txt
@@ -1,2 +1,3 @@
 idf_component_register(SRCS "spi_lcd_touch_example_main.c" "lvgl_demo_ui.c" "drivers/encoder.c" "drivers/display.c"
+                            "networking/http_handling.c"
                        INCLUDE_DIRS ".")
